+++ data/DownloadData.cmake
@@ -14,6 +14,7 @@ function(DownloadData)
         file(DOWNLOAD ${AWS_URL}/${DATA_FILE_CE} ${OUT} STATUS status)
         list(GET status 0 error_code)
         if( error_code )
+            message(STATUS "Error DOWNLOAD DATA, remove file")
             execute_process(
                     COMMAND
                     ${CMAKE_COMMAND} -E remove ${OUT}
@@ -21,12 +22,15 @@ function(DownloadData)
                     ${CMAKE_SOURCE_DIR}/data/ce
             )
         else()
+            message(STATUS "Extract file "${OUT})
+            message(STATUS "Working directory : "${CMAKE_SOURCE_DIR}"/data/ce/")
             execute_process(
                     COMMAND
                     ${CMAKE_COMMAND} -E tar xzf ${OUT}
-                    WORKING_DIRECTORY
-                    ${CMAKE_SOURCE_DIR}/data/ce
+                    WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/data/ce/
+                    RESULT_VARIABLE CMD_ERROR
             )
+            MESSAGE( STATUS "CMD_ERROR:" ${CMD_ERROR})
         endif()
     endif()
 endfunction()